pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gitlab-mr-auto-review"
version = "0.1.0"
description = "Auto review assigned GitLab MRs and send markdown reports"
requires-python = ">=3.9"
dependencies = [
  "python-gitlab>=4.5.0",
  "openai>=1.30.0",
  "requests>=2.31.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
